third_party/blink/web_tests/wpt_internal/long-animation-frame/loaf-ukm.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  const very_long_frame_duration = 250;
  const recorder = internals.initializeUKMRecorder();
  await new Promise(resolve => requestAnimationFrame(() => {
    const before = performance.now();
    document.querySelector("#log").innerHTML = "Update";
    while (performance.now() < (before + very_long_frame_duration)) {}
    requestAnimationFrame(resolve);
  }));

  let found = false;
  for (const load of ["BeforeLoad", "AfterLoad"]) {
    for (const blocking of ["NonBlocking", "BlockingFrame", "BlockingInput"]) {
      const entries = recorder.getMetrics(
        `PerformanceAPI.LongAnimationFrame.${load}.${blocking}`,
        ["Duration.Total", "DidPause"]);
      if (!entries.length || entries[0]["Duration.Total"] < very_long_frame_duration)
        continue;

      const metrics = entries[0];

      assert_equals(metrics["DidPause"], 0);
      found = true;
    }
  }

  assert_true(found);
}, "UKM recorder should receive LoAF entries");
</script>
<body>
  <div id="log"></div>
</body>
